pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.60.0",
    "groq>=1.0.0",
    "minsearch>=0.0.7",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "pydantic-ai>=1.50.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
]
